Move `MeterType` to Dummy Data seeder
